Remove MiniJSON
It worked well until stack nightly/new GHC started giving me
characters like '\190167' that I couldn't stuff into a JSON encoding
in a way that made my tests happy.

In practice, it wouldn't matter much, but getting it right and having
the tests be happy without telling it to avoid specific characters
seemed arbitrarily hard when aeson does the job OK.

I initially avoided aeson because a) it's a pretty big dependency and
b) MiniJSON was a lot faster.  But the dependency is already sunk
transitively by something, so I might as well just use it.  My ARM
builds are happening in docker now as well, so I'll just do the easy
thing.